refactored code
yes
