+++ RobotCode2020/src/main/java/frc/robot/OI.java
@@ -16,13 +16,6 @@ import edu.wpi.first.wpilibj.buttons.JoystickButton;
  * and command groups that allow control of the robot.
  */
 public class OI {
-  //// CREATING BUTTONS
-  // One type of button is a joystick button which is any button on a
-  //// joystick.
-  // You create one by telling it which joystick it's on and which button
-  // number it is.
-  // Joystick stick = new Joystick(port);
-  // Button button = new JoystickButton(stick, buttonNumber);
   public Joystick controller = new Joystick(RobotMap.CONTROLLER_PORT_ID);
   public Button abutton = new JoystickButton(controller, RobotMap.A_BUTTON_ID);
   public Button bbutton = new JoystickButton(controller, RobotMap.B_BUTTON_ID);
@@ -31,11 +24,27 @@ public class OI {
   public OI() {
     
   }
+
+  public double getAxis(int axis) {
+    return controller.getRawAxis(axis);
+  }
+}
+
+// documentation that was cluttering OI, but is useful for future reference
+
+//// CREATING BUTTONS
+  // One type of button is a joystick button which is any button on a
+  //// joystick.
+  // You create one by telling it which joystick it's on and which button
+  // number it is.
+  // Joystick stick = new Joystick(port);
+  // Button button = new JoystickButton(stick, buttonNumber);
+
   // There are a few additional built in buttons you can use. Additionally,
   // by subclassing Button you can create custom triggers and bind those to
   // commands the same as any other Button.
 
-  //// TRIGGERING COMMANDS WITH BUTTONS
+//// TRIGGERING COMMANDS WITH BUTTONS
   // Once you have a button, it's trivial to bind it to a button in one of
   // three ways:
 
@@ -50,8 +59,3 @@ public class OI {
   // Start the command when the button is released and let it run the command
   // until it is finished as determined by it's isFinished method.
   // button.whenReleased(new ExampleCommand());
-
-  public double getAxis(int axis) {
-    return controller.getRawAxis(axis);
-  }
-}
