+++ RobotCode2020/src/main/java/frc/robot/commands/ArmSet.java
@@ -11,10 +11,11 @@ import frc.robot.Robot;
 import frc.robot.RobotMap;
 import edu.wpi.first.wpilibj.command.Command;
 
+/**
+ * Rotates the arm based off joystick input.
+ */
 public class ArmSet extends Command {
   public ArmSet() {
-    // Use requires() here to declare subsystem dependencies
-    // eg. requires(chassis);
     requires(Robot.armMotor);
   }
