+++ RobotCode2020/src/main/java/frc/robot/commands/WinchWind.java
@@ -12,8 +12,6 @@ import frc.robot.Robot;
 
 public class WinchWind extends Command {
   public WinchWind() {
-    // Use requires() here to declare subsystem dependencies
-    // eg. requires(chassis);
     requires(Robot.climbDevice);
   }
