+++ RobotCode2020/src/main/java/frc/robot/commands/drivetrain/Drive.java
@@ -5,58 +5,46 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.drivetrain;
 
-import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.command.Command;
-import edu.wpi.first.wpilibj.controller.PIDController;
 import frc.robot.Robot;
-import frc.robot.utils.Limelight;
+import frc.robot.RobotMap;
 
-public class RotateToLimelightTarget extends Command {
-  private double P = 0.04;
-  private double I = 0.0;
-  private double D = 0.0055;
-  private PIDController pid = new PIDController(P, I, D);
-  private double commandStartTime;
-  public RotateToLimelightTarget() {
+public class Drive extends Command {
+  public Drive() {
     // Use requires() here to declare subsystem dependencies
     // eg. requires(chassis);
     requires(Robot.drivetrain);
-    pid.setTolerance(0.05);
-    commandStartTime = Timer.getFPGATimestamp();
   }
 
   // Called just before this Command runs the first time
   @Override
-  protected void initialize() {
-    Robot.drivetrain.calibrate();
-  }
+  protected void initialize() {}
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.drivetrain.drive(0, pid.calculate(Limelight.returnHorizontalOffset(), 0));
+    double speed =
+      -1 * Robot.m_oi.getAxis(RobotMap.JOYSTICK_DRIVE_FORWARDS_ID);
+    double rotation =
+      Robot.m_oi.getAxis(RobotMap.JOYSTICK_DRIVE_ROTATION_ID);
+    
+    Robot.drivetrain.drive(speed, rotation);
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    if (Timer.getFPGATimestamp() > commandStartTime + 2) {
-      return true;
-    } else {
-      return pid.atSetpoint();
-    }
+    return false;
   }
 
   // Called once after isFinished returns true
   @Override
-  protected void end() {
-  }
+  protected void end() {}
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
-  protected void interrupted() {
-  }
+  protected void interrupted() {}
 }
