+++ RobotCode2020/src/main/java/frc/robot/subsystems/Arm.java
@@ -16,7 +16,9 @@ import edu.wpi.first.wpilibj.controller.PIDController;
 
 import frc.robot.commands.ArmSet;
 
-/** Add your docs here. */
+/** 
+ * Arm lifting the intake
+ */
 public class Arm extends Subsystem {
   // Put methods for controlling this subsystem
   // here. Call these from Commands.
